Rewrite order parser
The order parser was pulling wrong locations when a portion of the
location name matched an abbreviation. Rewritten to use regexes so it's
easier to pull out full location names if specified.

Additionally, the fleet/army specification is optional. The unit type
will be pulled from the location if not specified