Consider multiple moves to cut support
When looking for only one move order to cut support there is a failure
when support is not cut because the cutting move is found to be the
location that the support is against. Consider all moves so that if any
move is not from the support toLocation the support is appropriately
cut.